backend/requirements.txt
wheel==0.46.1
fastapi==0.104.1
sqlmodel==0.0.24
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
openai==1.3.7
python-dotenv==1.0.0
httpx==0.25.2
SudachiPy==0.2.1
psycopg==3.2.9
